Fix tests for PetResource and PetResourceUnitTest
