+++ spring-petclinic-customers-service/src/test/java/org/springframework/samples/petclinic/customers/model/OwnerTest.java
@@ -2,7 +2,7 @@ package org.springframework.samples.petclinic.customers.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
+import java.util.Set;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
